- Consolidate file.ts below content folder - fileMimeType as optional attribute for Facebook channel -  Explain description of fileCaption as WhatsApp exclusive
